[DAT-859] Switch to new transfer protocol v3 (#12)
* Add dependencies required by proto from SDK

* Inject debug setup properties via local.properties file

* [DAT-746] Use java 11

* [DAT-725] Upgrade build.gradle dependencies

* Enable desugaring to support Java 11

* Upgrade cyface-utils to 3.0.1

* Upgade gradle to 7

* Use apache Valiadate

* [DAT-757] Handle paused deprecated measurement

* [DAT-747] Fix sync progress

* Upgrade dependencies

* [DAT-746] Fix lint errors

* Cleanup

* Fix error: app isnt installed

* Add gradle.properties.template

* Remove Apache Validate

* Read credentials from gradle.properties

* Upgrade dependencies and Gradle to 7

* Add maven repo for android-publish to settings.gradle

* Add support for DEPRECATED status

* Fix build

* Upgrade Cyface SDK, -camera service and energy settings

* Fix build after dependency upgrades

* [CY-5744] Fix no error shown on 401 error

* Remove backend dependency and add Serialization dependency 1.0.2

* Upgrade Cyface Utils and Energy Settings

* Upgrade Cyface Android Backend to 7.0.0

* Upgrade dependencies (Kotlin, Gradle Plugin, Annotation, Robolectr.)

* Upgrade Android Google map and location play services

* Undo fix to show UNAUTHORIZED toasts when app is in background

+++ measuring-client/src/main/java/de/cyface/app/ui/nav/view/MeasurementOverviewFragment.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Cyface GmbH
+ * Copyright 2017-2021 Cyface GmbH
  *
  * This file is part of the Cyface App for Android.
  *
@@ -70,7 +70,7 @@ import de.cyface.persistence.DefaultLocationCleaningStrategy;
 import de.cyface.persistence.DefaultPersistenceBehaviour;
 import de.cyface.persistence.PersistenceLayer;
 import de.cyface.persistence.model.Event;
-import de.cyface.persistence.model.GeoLocation;
+import de.cyface.persistence.model.ParcelableGeoLocation;
 import de.cyface.persistence.model.Measurement;
 import de.cyface.persistence.model.Modality;
 import de.cyface.persistence.model.Track;
@@ -82,7 +82,7 @@ import de.cyface.utils.Validate;
  *
  * @author Armin Schnabel
  * @author Klemens Muthmann
- * @version 4.5.4
+ * @version 4.5.5
  * @since 1.0.0
  */
 public class MeasurementOverviewFragment extends Fragment {
@@ -109,7 +109,7 @@ public class MeasurementOverviewFragment extends Fragment {
      */
     private boolean isEventsListShown = false;
     /**
-     * The {@link PersistenceLayer} required to retrieve the {@link Measurement} date from the {@link GeoLocation}s
+     * The {@link PersistenceLayer} required to retrieve the {@link Measurement} date from the {@link ParcelableGeoLocation}s
      */
     private PersistenceLayer<DefaultPersistenceBehaviour> persistenceLayer;
     /**
@@ -254,40 +254,39 @@ public class MeasurementOverviewFragment extends Fragment {
     public boolean onOptionsItemSelected(final MenuItem item) {
         final FragmentActivity fragmentActivity = getActivity();
         Validate.notNull(fragmentActivity);
-        switch (item.getItemId()) {
-            case R.id.export_menu_item:
-                // Permission requirements: https://developer.android.com/training/data-storage
-                final boolean requiresWritePermission = Build.VERSION.SDK_INT < Build.VERSION_CODES.Q;
-                final boolean missingPermissions = ContextCompat.checkSelfPermission(fragmentActivity,
-                        Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED
-                        || ContextCompat.checkSelfPermission(fragmentActivity,
-                                Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED;
-                if (requiresWritePermission && missingPermissions) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        requestPermissions(
-                                new String[] {Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                                        Manifest.permission.READ_EXTERNAL_STORAGE},
-                                Constants.PERMISSION_REQUEST_EXTERNAL_STORAGE_FOR_EXPORT);
-                    } else {
-                        Toast.makeText(fragmentActivity, fragmentActivity.getString(R.string.export_data_no_permission),
-                                Toast.LENGTH_LONG).show();
-                    }
+        if (item.getItemId() == R.id.export_menu_item) {
+            // Permission requirements: https://developer.android.com/training/data-storage
+            final boolean requiresWritePermission = Build.VERSION.SDK_INT < Build.VERSION_CODES.Q;
+            final boolean missingPermissions = ContextCompat.checkSelfPermission(fragmentActivity,
+                    Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED
+                    || ContextCompat.checkSelfPermission(fragmentActivity,
+                    Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED;
+            if (requiresWritePermission && missingPermissions) {
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                    requestPermissions(
+                            new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                                    Manifest.permission.READ_EXTERNAL_STORAGE},
+                            Constants.PERMISSION_REQUEST_EXTERNAL_STORAGE_FOR_EXPORT);
                 } else {
-                    new ExportTask(fragmentActivity).execute();
+                    Toast.makeText(fragmentActivity, fragmentActivity.getString(R.string.export_data_no_permission),
+                            Toast.LENGTH_LONG).show();
                 }
-                return true;
-            case R.id.select_all_item:
-                selectAllItems();
-                return true;
-            case R.id.delete_measurement_item:
-                if (isEventsListShown) {
-                    deleteSelectedEvents(fragmentActivity, eventDataList.getListView());
-                } else {
-                    deleteSelectedMeasurements(fragmentActivity, measurementDataList.getListView());
-                }
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+            } else {
+                new ExportTask(fragmentActivity).execute();
+            }
+            return true;
+        } else if (item.getItemId() == R.id.select_all_item) {
+            selectAllItems();
+            return true;
+        } else if (item.getItemId() == R.id.delete_measurement_item) {
+            if (isEventsListShown) {
+                deleteSelectedEvents(fragmentActivity, eventDataList.getListView());
+            } else {
+                deleteSelectedMeasurements(fragmentActivity, measurementDataList.getListView());
+            }
+            return true;
+        } else {
+            return super.onOptionsItemSelected(item);
         }
     }
 
@@ -451,7 +450,7 @@ public class MeasurementOverviewFragment extends Fragment {
             markerLocation.setLongitude(markerPosition.longitude);
 
             // Load GeoLocations
-            final List<GeoLocation> geoLocations = new ArrayList<>();
+            final List<ParcelableGeoLocation> geoLocations = new ArrayList<>();
             try {
                 List<Track> tracks = persistenceLayer.loadTracks(measurementId, new DefaultLocationCleaningStrategy());
                 for (final Track track : tracks) {
@@ -463,8 +462,8 @@ public class MeasurementOverviewFragment extends Fragment {
 
             // Search for the nearest GeoLocation
             Double minDistance = null;
-            GeoLocation nearestGeoLocation = null;
-            for (final GeoLocation geoLocation : geoLocations) {
+            ParcelableGeoLocation nearestGeoLocation = null;
+            for (final ParcelableGeoLocation geoLocation : geoLocations) {
                 final Location location = new Location("geoLocation");
                 location.setLatitude(geoLocation.getLat());
                 location.setLongitude(geoLocation.getLon());
